Add telescope
Finalize the move from the old vimrc system by adding telescope.nvim
for some nice reference finding and other stuff.

Fix the ranger issue with the keybinds by setting the variables before
it's loaded.

Make the Lazy window not have the trailing white space highlight.

Map some of the old ALE functions to the built in LSP ones that work
the same way.